+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2024-09-26
+
+### Added
+
+- Candl method: `getXLabels()` to get all the xLabels.
+- Candl method: `setXLabelsStyle()` to set the style of all xLabels at once.
+
 ## [0.5.0] - 2024-09-26
 
 ### Added